+++ app/src/main/res/layout/track_item.xml
@@ -9,12 +9,15 @@
 
     <ImageView
         android:id="@+id/albums_cover"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
         android:layout_width="45dp"
         android:layout_height="45dp"
+        android:layout_marginLeft="13dp"
         android:layout_marginTop="8dp"
-        android:layout_marginStart="13dp" />
+        android:layout_marginBottom="8dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        />
 
     <TextView
         android:id="@+id/track_name"
@@ -22,13 +25,11 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
         android:layout_marginTop="14dp"
-        android:layout_marginEnd="16dp"
-        android:layout_marginBottom="1dp"
+        android:layout_marginEnd="8dp"
         android:ellipsize="end"
         android:maxLines="1"
         android:singleLine="true"
         android:theme="@style/rvTextStyle"
-        app:layout_constraintBottom_toTopOf="@+id/linearLayout"
         app:layout_constraintEnd_toStartOf="@+id/vector_button"
         app:layout_constraintStart_toEndOf="@+id/albums_cover"
         app:layout_constraintTop_toTopOf="parent" />
@@ -49,7 +50,7 @@
             android:ellipsize="end"
             android:maxLines="1"
             android:singleLine="true"
-            android:textSize="11sp"
+            android:textSize="@dimen/text_size_item"
             android:theme="@style/rvLilTextStyle" />
 
         <ImageView
@@ -71,15 +72,17 @@
             android:theme="@style/rvLilTextStyle"
             />
     </LinearLayout>
+
     <ImageView
         android:id="@+id/vector_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        android:src="@drawable/forward"
         android:layout_marginTop="18dp"
         android:layout_marginEnd="12dp"
+        android:layout_marginBottom="19dp"
+        android:src="@drawable/forward"
         android:theme="@style/rvForwardStyle"
-        />
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
